+++ CHANGELOG.md
@@ -19,6 +19,7 @@ Please view this file on the master branch, on stable branches it's out of date.
 - information texts in service ordering page (@goreck888)
 - `service_type` to `order_type` in service and offer (@goreck888)
 - `tutorial_url` to `training_information_url` in service model (@goreck888)
+- Service `title` to `name` (@goreck888)
 
 ### Deprecated